requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pymupdf>=1.22.0
pypdf2>=3.0.0
langchain>=0.1.0
faiss-cpu>=1.7.4
# sentence-transformers>=2.2.2
google-generativeai>=0.3.0
openpyxl>=3.1.2
tqdm>=4.66.0
rank_bm25>=0.2.2
